+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [1.5.0]
+ * modify styles.dart
+ * the modification conflicts with the method of the camera component
 ## [1.3.6]
  * remove TZImagePickerController and picture_library
  * remove openImagePicker